moved get_vendor_code to utils.py
