+++ tests/test_adapters.py
@@ -5,7 +5,6 @@ import pytest
 from record_validator.adapters import (
     get_adapter,
     get_record_type,
-    get_vendor_code,
     pattern_match,
     tag_discriminator,
     AuxOtherFields,
@@ -186,47 +185,6 @@ def test_get_record_type_dict(stub_record):
     assert (
         get_record_type(fields=record_dict["fields"], vendor_code="EVP") == "monograph"
     )
-
-
-def test_get_vendor_code_evp(stub_record):
-    assert get_vendor_code(fields=stub_record.fields) == "EVP"
-
-
-def test_get_vendor_code_auxam(stub_aux_other_record):
-    assert get_vendor_code(fields=stub_aux_other_record.fields) == "AUXAM"
-
-
-def test_get_vendor_code_leila(stub_leila_monograph):
-    assert get_vendor_code(fields=stub_leila_monograph.fields) == "LEILA"
-
-
-def test_get_vendor_code_none():
-    assert get_vendor_code(fields=["foo"]) is None
-
-
-def test_get_vendor_code_bad_input(stub_record):
-    stub_record["901"].add_subfield("a", "foo")
-    assert get_vendor_code(fields=stub_record.fields) is None
-
-
-def test_get_vendor_code_dict():
-    fields = [
-        {"tag": "001", "value": "on1381158740"},
-        {
-            "tag": "901",
-            "ind1": " ",
-            "ind2": " ",
-            "subfields": [
-                {"a": "EVP"},
-            ],
-        },
-    ]
-    assert get_vendor_code(fields=fields) == "EVP"
-
-
-def test_get_vendor_code_pymarc_dict(stub_record):
-    stub_record_dict = stub_record.as_dict()
-    assert get_vendor_code(fields=stub_record_dict["fields"]) == "EVP"
 
 
 @pytest.mark.parametrize(
